RA-884 Correcting Automated Tests

+++ ui-tests/src/test/java/org/openmrs/reference/PasswordMinimumLengthTest.java
@@ -31,7 +31,6 @@ public class PasswordMinimumLengthTest extends TestBase {
         homePage.goToAdministration();
     }
 
-    @Ignore
     @Test
     public void passwordMinimumLengthTest() throws Exception {
         settingPage.clickOnSetting();
@@ -41,23 +40,29 @@ public class PasswordMinimumLengthTest extends TestBase {
         headerPage.clickOnHomeLink();
         homePage.goToAdministration();
         administrationPage.clickOnManageUsers();
-        manageUserPage.clickOnAddUser();
-        manageUserPage.createNewPerson();
-        manageUserPage.enterUserMale("doctor", "House", "dr_house", "house");
-        manageUserPage.chooseRole();
-        manageUserPage.saveUser();
-        assertTrue(driver.getPageSource().contains("Password should be 6 characters long"));
-        manageUserPage.changePassword("House1");
-        manageUserPage.saveUser();
-        manageUserPage.findUser("dr_house");
-        manageUserPage.deleteUser();
-        headerPage.clickOnHomeLink();
-        homePage.goToAdministration();
-        settingPage.clickOnSetting();
-        settingPage.clickOnSecurity();
-        settingPage.enterLength("8");
-        assertTrue(driver.getPageSource().contains("Global properties saved"));
-    }
+        manageUserPage.checkUser("dr_house");
+        if (manageUserPage.userExist("dr_house")) {
+            manageUserPage.clickOnUser();
+            manageUserPage.deleteUser();
+        }
+            manageUserPage.clickOnAddUser();
+            manageUserPage.createNewPerson();
+            manageUserPage.enterUserMale("doctor", "House", "dr_house", "house");
+            manageUserPage.chooseRole();
+            manageUserPage.saveUser();
+            assertTrue(driver.getPageSource().contains("Password should be 6 characters long"));
+            manageUserPage.changePassword("House1");
+            manageUserPage.saveUser();
+            manageUserPage.findUser("dr_house");
+            manageUserPage.deleteUser();
+            headerPage.clickOnHomeLink();
+            homePage.goToAdministration();
+            settingPage.clickOnSetting();
+            settingPage.clickOnSecurity();
+            settingPage.enterLength("8");
+            settingPage.waitForMessage();
+            assertTrue(driver.getPageSource().contains("Global properties saved"));
+        }
 
     @After
     public void tearDown() throws Exception {
